Create jest project for testing internals
Creates a jest project for testing APIs that aren't part of the public
API. Any `*-test.js` file in a `__tests__` directory will be considered
an "internal" test and can be colocated with the relevant internal
module.
